requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
numpy>=1.24.0
Pillow>=10.0.0
segyio>=1.9.0